requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
boto3==1.34.158
botocore==1.34.158
certifi==2024.7.4
click==8.1.7
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
jiter==0.5.0
jmespath==1.0.1
jsonlines==3.1.0
nest-asyncio==1.6.0
numpy==2.0.1
openai==1.40.2
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pydantic==1.10.8
pydantic_core==2.20.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
